requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Compress==1.15
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.37.1
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zstandard==0.23.0
